[TECH-521] Move push and pop to build command
branch: feature/TECH-521-simplified-remote-artifact-caching
