+++ tests/cli/test_resources/build_help_text.txt
@@ -12,5 +12,7 @@ Options:
 Commands:
   clean    Clean MPyL metadata in `.mpyl` folders
   jenkins  Run a multi branch pipeline build on Jenkins
+  pull     Pull build artifacts from remote artifact repository
+  push     Push build artifacts to remote artifact repository
   run      Run an MPyL build
   status   The status of the current local branch from MPyL's perspective
