+++ tests/cli/test_resources/main_help_text.txt
@@ -6,9 +6,8 @@ Options:
   --help  Show this message and exit.
 
 Commands:
-  artifacts  Manage remote build artifacts
-  build      Pipeline build commands
-  health     Health check
-  projects   Commands related to MPyL project configurations (project.yml)
-  repo       Manage CVS (git) repositories
-  version    Version information
+  build     Pipeline build commands
+  health    Health check
+  projects  Commands related to MPyL project configurations (project.yml)
+  repo      Manage CVS (git) repositories
+  version   Version information
